Push AnnotatedPipeline into //compiler/annotation
