+++ compiler/annotation/annotated_functions.rs
@@ -6,7 +6,7 @@
 
 use encoding::graph::definition::definition_key::DefinitionKey;
 use ir::program::{function::Function, function_signature::FunctionIDAPI};
-use crate::inference::type_annotations::FunctionAnnotations;
+use crate::annotation::type_annotations::FunctionAnnotations;
 
 
 /// Indexed by Function ID
