+++ compiler/annotation/mod.rs
@@ -5,14 +5,29 @@
  */
 
 use concept::error::ConceptReadError;
+use encoding::value::value_type::ValueTypeCategory;
 use error::typedb_error;
+use crate::expression::ExpressionCompileError;
 
 pub mod match_inference;
 pub mod type_annotations;
 pub mod type_inference;
 mod type_seeder;
 pub mod annotated_functions;
-pub mod annotated_program;
+pub mod pipeline;
+
+typedb_error!(
+    pub AnnotationError(component = "Query annotation", prefix = "QUA") {
+        TypeInference(0, "Type inference error while compiling query annotations.", ( typedb_source : TypeInferenceError )),
+        PreambleTypeInference(1, "Type inference error while compiling query premable functions.", ( typedb_source : FunctionTypeInferenceError )),
+        ExpressionCompilation(2, "Error inferring correct expression types.", ( source : ExpressionCompileError )),
+        CouldNotDetermineValueTypeForReducerInput(15, "The value-type for the reducer input variable '{variable}' could not be determined.", variable: String),
+        ReducerInputVariableDidNotHaveSingleValueType(16, "The reducer input variable '{variable}' had multiple value-types.", variable: String),
+        UnsupportedValueTypeForReducer(17, "The input variable to the reducer'{reducer}({variable})' reducer had an unsupported value-type: '{value_type}'", reducer: String, variable: String, value_type: ValueTypeCategory),
+        UncomparableValueTypesForSortVariable(18, "The sort variable '{variable}' could return incomparable value-types '{category1}' & '{category2}'.", variable: String, category1: ValueTypeCategory, category2: ValueTypeCategory),
+    }
+);
+
 
 
 typedb_error!(
@@ -90,7 +105,7 @@ pub mod tests {
     };
     use storage::{durability_client::WALClient, MVCCStorage};
     use test_utils::{create_tmp_dir, init_logging, TempDir};
-    use crate::inference::match_inference::{NestedTypeInferenceGraphDisjunction, TypeInferenceEdge, TypeInferenceGraph};
+    use crate::annotation::match_inference::{NestedTypeInferenceGraphDisjunction, TypeInferenceEdge, TypeInferenceGraph};
 
     impl<'this> PartialEq<Self> for TypeInferenceEdge<'this> {
         fn eq(&self, other: &Self) -> bool {
