+++ compiler/match_/planner/plan.rs
@@ -27,7 +27,7 @@ use itertools::Itertools;
 
 use crate::{
     expression::compiled_expression::CompiledExpression,
-    inference::type_annotations::TypeAnnotations,
+    annotation::type_annotations::TypeAnnotations,
     match_::{
         instructions::{
             thing::{
